tests/wpt/web-platform-tests/infrastructure/expected-fail/uncaught-exception-single-test.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Uncaught exception in single-page test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({ single_test: true });
throw new Error("error outside any setup or test");
</script>
