LayoutTests/imported/w3c/web-platform-tests/infrastructure/expected-fail/uncaught-exception-single-test-expected.txt

CONSOLE MESSAGE: Error: error outside any setup or test

FAIL Uncaught exception in single-page test Error: error outside any setup or test
